Move evil_genius_labs coordinator to separate module
